### 2019-09-16 #1739 @whedon
[PRE REVIEW]: Economiccomplexity: Methods Used in the Economic Complexity Literature
**Submitting author:** @pachamaltese (<a href="http://orcid.org/0000-0003-1017-7574">Mauricio Vargas</a>)
**Repository:** <a href="https://github.com/pachamaltese/economiccomplexity" target ="_blank">https://github.com/pachamaltese/economiccomplexity</a>
**Version:** 0.2.1
