### 2019-09-16 #1739 @kthyng
Hi @pachamaltese! Can you explain a little bit more about the research application of your software? Please keep in mind in particular the [author guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Thanks!
