### 2019-09-17 #1739 @pachamaltese
Hi @kthyng! sure, this package is the only I've seen on CRAN, and amongst R packages, to explore bipartite relations used in Economic Complexity. This takes all the equations (i.e LaTeX equations) from The Atlas of Economic Complexity and related articles, to R functions that, in my opinion, are easy to use and may contribute to ease reproducibility in Social Sciences.
I've read that the functions from this packages are also used in Economic Geography and may be useful to explore different relations in Media Sociology such as twitter users and twitter contents.
I can say that this introduces in an easy+fast way the numerical methods from Economic Complexity into the R-world.
