### 2019-09-17 fec9c11
fixed ubiquity with tbl = T and any column names
