+++ inst/REFERENCES.bib
@@ -140,17 +140,17 @@
 }
 
 @article{scientific2014,
-    author = {Cimini, Giulio AND Gabrielli, Andrea AND Sylos Labini, Francesco},
-    journal = {PLOS ONE},
-    publisher = {Public Library of Science},
-    title = {The Scientific Competitiveness of Nations},
-    year = {2014},
-    month = {12},
-    volume = {9},
-    url = {https://doi.org/10.1371/journal.pone.0113470},
-    pages = {1-11},
-    number = {12},
-    doi = {10.1371/journal.pone.0113470}
+  author = {Cimini, Giulio and Gabrielli, Andrea and Sylos Labini, Francesco},
+  journal = {PLOS ONE},
+  publisher = {Public Library of Science},
+  title = {The Scientific Competitiveness of Nations},
+  year = {2014},
+  month = {12},
+  volume = {9},
+  url = {https://doi.org/10.1371/journal.pone.0113470},
+  pages = {1-11},
+  number = {12},
+  doi = {10.1371/journal.pone.0113470}
 }
 
 @article{measuring2014,
