### 2019-09-17 #1739 @pachamaltese
fixing wrong bib entry from plos one
