### 2019-09-21 #1739 @oliviaguest
👋 Hey @mdsumner — would you be interested in reviewing this submission for JOSS? 😸
