### 2019-10-09 18788fb
small path to single proximity calculation
