### 2019-10-17 #1739 @oliviaguest
@economistgame — would you be interested in reviewing this submission for JOSS? :blush:
