### 2019-10-17 #1739 @pachamaltese
> The header comment in this issue does request reviewer suggestions from the author.
>
> > @pachamaltese if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.

sure, based on networks/graphs/economy

* MatthewSmith430 (visualisation, international trade, network analysis)
* paulgovan
* JohnCoene
* Nick-gauthier
