### 2019-11-05 #1739 @JohnCoene
@oliviaguest @pachamaltese I apologise for the belated reply, the notifications and emails totally escaped me. I'll be more than happy to review the package, it looks very interesting!
