### 2019-11-05 #1739 @rickecon
@oliviaguest, I saw that @JohnCoene responded. In the future, try inviting the following two individuals to review JOSS economics manuscripts. Both are open source modelers and both would likely be interested in contributing to this community and to JOSS.

@jdebacker, Jason DeBacker, Associate Professor, Department of Economics, University of South Carolina ([https://www.jasondebacker.com/](https://www.aei.org/profile/matthew-h-jensen/))

@MattHJensen, Matt Jensen, Director of the Open Source Policy Center at the American Enterprise Institute ([https://www.aei.org/profile/matthew-h-jensen/](https://www.aei.org/profile/matthew-h-jensen/))
