### 2019-11-05 #1866 @whedon
[REVIEW]: Economiccomplexity: Methods Used in the Economic Complexity Literature
**Submitting author:** @pachamaltese (<a href="http://orcid.org/0000-0003-1017-7574">Mauricio Vargas</a>)
**Repository:** <a href="https://github.com/pachamaltese/economiccomplexity" target ="_blank">https://github.com/pachamaltese/economiccomplexity</a>
**Version:** 0.2.1

- [NA] **State of the field:** Do the authors describe how this software compares to other commonly-used packages? -- there seems to be no other comparable software
- [x] **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? - the writing could perhaps do with one final, minor edit
