### 2019-11-10 pachamaltese/economiccomplexity#8 @JohnCoene
Function Description
Functions are missing `@description` that explains in more detail what each function do, e.g.: formula used to compute [revealed comparative advantage](https://en.wikipedia.org/wiki/Revealed_comparative_advantage)
