### 2019-11-19 #1866 @JohnCoene
@oliviaguest I have already started and submitted an [issue on the repository](https://github.com/pachamaltese/economiccomplexity/issues/8) itself, will finish review by end of week.
