### 2019-11-19 #1866 @pachamaltese
yes, I'm improving the documentation after those initial comments
