+++ R/productivity_levels.R
@@ -43,7 +43,6 @@
 #'   v2 = "value",
 #'   tbl = TRUE
 #' )
-#'
 #' @references
 #' For more information on prody and its applications see:
 #'
