+++ R/revealed_comparative_advantage.R
@@ -1,23 +1,28 @@
 #' Revealed Comparative Advantage (RCA)
 #'
 #' @export
-#' @param data tibble/data.frame in long format, it must contain the
-#' columns c (character/factor), p (character/factor) and export
-#' v (numeric)
-#' @param c string to indicate the column that contains exporting
-#' countries (e.g. "reporter_iso")
-#' @param p string to indicate the column that contains exported ps
-#' (e.g. "p_code")
-#' @param v string to indicate the column that contains traded vs
-#' (e.g. "trade_v_usd")
-#' @param discrete when set to TRUE it will convert all the Revealed Comparative
-#' Advantage vs
-#' to zero or one based on the cutoff v (default set to TRUE)
-#' @param cutoff when set to TRUE all the vs lower than the specified cutoff
-#' will be converted to zero and to one in other case, numeric (default set
-#' to 1)
-#' @param tbl when set to TRUE the output will be a tibble instead of a
-#' matrix (default set to FALSE)
+#'
+#' @description Given a \eqn{C\times P} matrix (C for "countries"
+#' and P for "products") or an equivalent 3-columns data.frame with exported
+#' values (X) as input, this function computes RCA metric.
+#' The equation used in this function is:
+#' \deqn{RCA_{cp} = \frac{X_{cp}}{\sum_c X_{cp}} /
+#' \frac{\sum_p X_{cp}}{\sum_{c}\sum_{p} X_{cp}}}
+#'
+#' @param data matrix or data.frame with traded values
+#' @param country column containing countries (applies only if d is a
+#' data.frame)
+#' @param product column containing products (applies only if d is a
+#' data.frame)
+#' @param value column containing traded values (applies only if d is a
+#' data.frame)
+#' @param discrete convert to one all the values above a cutoff and zero
+#' otherwise (by default is TRUE)
+#' @param cutoff all the values below the specified will be converted to zero
+#' (by default is 1)
+#' @param tbl by default the output is a data.frame unless this is changed to
+#' FALSE, in which case the output is a matrix
+#'
 #' @importFrom magrittr %>%
 #' @importFrom dplyr select group_by ungroup mutate summarise matches rename
 #' pull as_tibble if_else
@@ -26,11 +31,7 @@
 #' @importFrom rlang sym syms :=
 #'
 #' @examples
-#' ec_rca(
-#'   data = ec_trade_1962,
-#'   tbl = TRUE
-#' )
-#'
+#' ec_rca(ec_trade_1962)
 #' @references
 #' For more information on revealed comparative advantage and its uses see:
 #'
@@ -38,23 +39,21 @@
 #'
 #' @keywords functions
 
-ec_rca <- function(data = NULL,
-                   c = "country",
-                   p = "product",
-                   v = "value",
-                   cutoff = 1,
+ec_rca <- function(data,
+                   country = "country",
+                   product = "product",
+                   value = "value",
                    discrete = TRUE,
-                   tbl = FALSE) {
+                   cutoff = 1,
+                   tbl = TRUE) {
   # sanity checks ----
-  if (all(class(data) %in% c(
-    "data.frame", "matrix", "dgeMatrix",
-    "dsCMatrix", "dgCMatrix"
-  ) == FALSE)) {
[…]
