+++ R/complexity_measures.R
@@ -2,28 +2,34 @@
 #'
 #' @export
 #'
-#' @param rca matrix or tibble/data.frame (e.g. the
-#' output of \code{revealed_comparative_advantage()}).
-#' If the input is a matrix it must be a zero/one matrix with countries in rows
-#' and ps in columns.
-#' If the input is a tibble/data.frame it must contain at least three columns
-#' with countries, ps and vs.
-#' @param c string to indicate the column that contains exporting
-#' countries in rca (set to "country" by default)
-#' @param p string to indicate the column that contains exported ps
-#' in rca (set to "product" by default)
-#' @param v string to indicate the column that contains traded vs in
-#' rca (set to "value" by default)
-#' @param method string to indicate to use one of these methods: reflections,
-#' eigenvalues or fitness (set to "fitness" by default)
-#' @param atlas logical v to remove the countries not ranked in The
-#' Atlas of Economic Complexity (set to FALSE by default)
-#' @param iterations number of iterations to use in the reflections method
-#' (set to 20 by default)
+#' @description Given a \eqn{C\times P} matrix (C for "countries"
+#' and P for "products") or an equivalent 3-columns data.frame with RCA
+#' values as input, this function computes Country/Product Complexity Indices
+#' and Diversity/Ubiquity metrics.
+#' The equations used in this function are the generalized fitness
+#' \eqn{F(\gamma)} and complexity \eqn{Q(\gamma)} defined by:
+#' \deqn{\tilde{F}^{(n)}(\gamma) = \sum_{p} R_{cp}Q_{p}^{n-1}}
+#' \deqn{\tilde{Q}^{(n)}(\gamma) =
+#' \left[\sum_{c} R_{cp}(F_{c}^{n-1})^{-\gamma}\right]^{-1/\gamma}}
+#' Altogether with equivalent formulations from other methods.
+#'
+#' @param rca matrix or data.frame with RCA values
+#' @param country column containing countries (applies only if d is a
+#' data.frame)
+#' @param product column containing products (applies only if d is a
+#' data.frame)
+#' @param value column containing traded values (applies only if d is a
+#' data.frame)
+#' @param method which method to use (by default is "fitness", it can also
+#' be "reflections" or "eigenvalues")
+#' @param iterations number of iterations to use in the fitness, reflections
+#' and indirectly in eigenvalues method (by default is 20)
 #' @param extremality numeric coefficient to use in the fitness method
-#' (set to 1 by default)
-#' @param tbl logical v to use tibble output instead of a matrix
-#' output (set to FALSE by default)
+#' (by default is 1)
+#' @param atlas remove the countries not ranked in The Atlas of Economic
+#' Complexity (by default is FALSE)
+#' @param tbl by default the output is a data.frame unless this is changed to
+#' FALSE, in which case the output is a matrix
 #'
 #' @importFrom magrittr %>%
 #' @importFrom dplyr select mutate arrange pull rename
@@ -33,36 +39,36 @@
 #' @importFrom rlang sym
 #'
 #' @examples
-#' ec_complexity_measures(
-#'   rca = ec_output_demo$rca_tbl,
-#'   tbl = TRUE
-#' )
+#' ec_complexity_measures(ec_output_demo$rca_tbl)
 #'
 #' @references
-#' For more information on complexity measures, indices and its applications
-#' see:
-#'
-#' \insertRef{atlas2014}{economiccomplexity}
+#' For more information on complexity measures see:
 #'
 #' \insertRef{measuringcomplexity2015}{economiccomplexity}
 #'
[…]
