+++ man/ec_networks.Rd
@@ -41,7 +41,6 @@ ec_networks(
   pp = ec_output_demo$proximity_tbl$proximity_p,
   tbl = TRUE
 )
-
 }
 \references{
 For more information on networks such as the product space and its

[…]
