+++ R/proximity.R
@@ -2,40 +2,25 @@
 #'
 #' @export
 #'
-#' @param rca matrix or tibble/data.frame in long format (e.g. the output of
-#' revealed_comparative_advantage()). If it is a matrix it must be a zero/one
-#' matrix with countries in the row names and ps in the column names.
-#' If rca is a tibble/data.frame it must contain the columns c (character/
-#' factor), p (character/factor) and discrete RCA (integer)
-#' @param c string to indicate the column that contains exporting
-#' countries (default set to "country" that is the output of
-#' revealed_comparative_advantage(), applies only if rca is a data.frame)
-#' @param p string to indicate the column that contains exported ps
-#' (default set to "product" that is the output of
-#' revealed_comparative_advantage(), applies only if rca is a data.frame)
-#' @param v string to indicate the column that contains RCA vs (default
-#' set to "value" that is the output of revealed_comparative_advantage(),
-#' applies only if rca is a data.frame)
-#' @param d numeric vector or tibble/data.frame containing diversity
-#' measures (e.g. \code{d} from \code{complexity_measures()})
-#' @param u numeric vector or tibble/data.frame containing ubiquity
-#' measures (e.g. \code{u} from \code{complexity_measures()})
-#' @param d_c string to indicate the column that contains
-#' countries in "diversity" (default set to "country" that is the output of
-#' economic_complexity_measures())
-#' @param d_v string to indicate the column that contains
-#' values in "diversity" (default set to "value" that is the output of
-#' economic_complexity_measures())
-#' @param u_p string to indicate the column that contains
-#' products in "ubiquity" (default set to "product" that is the output of
-#' economic_complexity_measures())
-#' @param u_v string to indicate the column that contains
-#' values in "ubiquity" (default set to "value" that is the output of
-#' economic_complexity_measures())
-#' @param tbl when set to TRUE the output will be a tibble instead of a
-#' matrix (default set to FALSE)
-#' @param compute by default set to "both", it can also be "country" or
-#' "product"
+#' @param rca matrix or data.frame with RCA values
+#' @param diversity matrix or data.frame with diversity values
+#' @param ubiquity matrix or data.frame with ubiquity values
+#' @param country_r column containing countries (applies only if d is a
+#' data.frame)
+#' @param product_r column containing products (applies only if d is a
+#' data.frame)
+#' @param value_r column containing traded values (applies only if d is a
+#' data.frame)
+#' @param country_d column containing countries (applies only if d is a
+#' data.frame)
+#' @param value_d column containing values (applies only if d is a
+#' data.frame)
+#' @param product_u column containing products (applies only if d is a
+#' data.frame)
+#' @param value_u column containing values (applies only if d is a
+#' data.frame)
+#' @param tbl TRUE (default) returns a data.frame and FALSE returns a matrix
+#' @param compute "country", "product" or "both" (default) matrices
 #'
 #' @importFrom magrittr %>%
 #' @importFrom dplyr select filter mutate pull
@@ -48,8 +33,7 @@
 #' ec_proximity(
 #'   rca = ec_output_demo$rca_tbl,
 #'   d = ec_output_demo$complexity_measures_tbl$diversity,
[…]
