+++ R/networks.R
@@ -1,25 +1,19 @@
-#' Networks
+#' @title Country-Product bipartite network projections
 #'
-#' @export
-#'
-#' @param pc matrix or tibble/data.frame, if d is a tibble/
+#' @param proximity_c matrix or tibble/data.frame, if d is a tibble/
 #' data.frame it must contain the columns from (character/factor), to
 #' (character/factor) and value (numeric), if it is a matrix it must be a
 #' numeric matrix with countries in row names and column names
-#' @param pp matrix or tibble/data.frame, if d is a
+#' @param proximity_p matrix or tibble/data.frame, if d is a
 #' tibble/data.frame it must contain the columns from (character/factor), to
 #' (character/factor) and value (numeric), if it is a matrix it must be a
 #' numeric matrix with products in row names and column names
-#' @param cutoff_c all the values below the specified
-#' cutoff_c will be converted to 0 and excluded from the countries
-#' network (default set to 0.2)
-#' @param cutoff_p all the values below the specified
-#' cutoff_p will be converted to 0 and excluded from the products network
-#' (default set to 0.4)
-#' @param tbl when set to TRUE the output will be a tibble instead of a
-#' graph (default set to FALSE)
-#' @param compute by default set to "both", it can also be "country" or
-#' "product"
+#' @param cutoff_c all the links with a proximity below this value will be
+#' removed from the country projection (by default is 0.2)
+#' @param cutoff_p all the links with a proximity below this value will be
+#' removed from the product projection (by default is 0.4)
+#' @param tbl TRUE (default) returns a data.frame and FALSE returns a matrix
+#' @param compute "country", "product" or "both" (default) projections
 #'
 #' @importFrom magrittr %>%
 #' @importFrom dplyr as_tibble filter mutate bind_rows
@@ -27,37 +21,34 @@
 #' @importFrom igraph graph_from_data_frame mst as_data_frame simplify
 #' @importFrom rlang sym
 #'
+#' @return A list with two data frames or two matrices.
+#'
 #' @examples
 #' ec_networks(
-#'   pc = ec_output_demo$proximity_tbl$proximity_c,
-#'   pp = ec_output_demo$proximity_tbl$proximity_p,
-#'   tbl = TRUE
+#'   ec_output_demo$proximity_tbl$proximity_c
+#'   ec_output_demo$proximity_tbl$proximity_p
 #' )
+#'
 #' @references
-#' For more information on networks such as the product space and its
-#' applications see:
+#' For more information about bipartite networks projections see:
+#' aproximity_plications see:
 #'
 #' \insertRef{atlas2014}{economiccomplexity}
 #'
 #' @keywords functions
 
-ec_networks <- function(pc,
-                        pp,
+projections <- function(proximity_c,
+                        proximity_p,
                         cutoff_c = 0.2,
                         cutoff_p = 0.4,
-                        tbl = FALSE,
+                        tbl = TRUE,
                         compute = "both") {
   # sanity checks ----
-  if (all(class(pc) %in% c(
-    "data.frame", "matrix",
-    "dgeMatrix", "dgCMatrix", "dsCMatrix"
-  ) == FALSE) &
-    all(class(pp) %in% c(
[…]
