+++ vignettes/basic-usage.R
@@ -3,7 +3,6 @@ knitr::opts_chunk$set(eval = TRUE, message = FALSE, warning = FALSE)
 
 ## ------------------------------------------------------------------------
 library(economiccomplexity)
-
 ec_trade_1962
 
 ## ------------------------------------------------------------------------
@@ -11,28 +10,12 @@ r <- rca(ec_trade_1962)
 r
 
 ## ------------------------------------------------------------------------
-r2 <- rca(ec_trade_1962, tbl = F)
-r2[1:5,1:5] # 5x5 preview
-
-## ------------------------------------------------------------------------
-r3 <- rca(ec_trade_1962, discrete = F)
-r3
-
-## ------------------------------------------------------------------------
-r4 <- rca(ec_trade_1962, discrete = F, tbl = F)
-r4[1:5,1:5] # 5x3 preview
+r2 <- rca(ec_trade_1962, discrete = F)
+r2
 
 ## ------------------------------------------------------------------------
 cf <- complexity(r, method = "fitness")
 cf
-
-## ------------------------------------------------------------------------
-cr <- complexity(r, method = "reflections")
-cr
-
-## ------------------------------------------------------------------------
-ce <- complexity(r, method = "eigenvalues")
-ce
 
 ## ------------------------------------------------------------------------
 p <- proximity(r, cf$country_diversity, cf$product_ubiquity)
