+++ man/projections.Rd
@@ -1,0 +1,50 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/networks.R
+\name{projections}
+\alias{projections}
+\title{Country-Product bipartite network projections}
+\usage{
+projections(proximity_c, proximity_p, cutoff_c = 0.2, cutoff_p = 0.4,
+  tbl = TRUE, compute = "both")
+}
+\arguments{
+\item{proximity_c}{matrix or tibble/data.frame, if d is a tibble/
+data.frame it must contain the columns from (character/factor), to
+(character/factor) and value (numeric), if it is a matrix it must be a
+numeric matrix with countries in row names and column names}
+
+\item{proximity_p}{matrix or tibble/data.frame, if d is a
+tibble/data.frame it must contain the columns from (character/factor), to
+(character/factor) and value (numeric), if it is a matrix it must be a
+numeric matrix with products in row names and column names}
+
+\item{cutoff_c}{all the links with a proximity below this value will be
+removed from the country projection (by default is 0.2)}
+
+\item{cutoff_p}{all the links with a proximity below this value will be
+removed from the product projection (by default is 0.4)}
+
+\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}
+
+\item{compute}{"country", "product" or "both" (default) projections}
+}
+\value{
+A list with two data frames or two matrices.
+}
+\description{
+Country-Product bipartite network projections
+}
+\examples{
+ec_networks(
+  ec_output_demo$proximity_tbl$proximity_c
+  ec_output_demo$proximity_tbl$proximity_p
+)
+
+}
+\references{
+For more information about bipartite networks projections see:
+aproximity_plications see:
+
+\insertRef{atlas2014}{economiccomplexity}
+}
+\keyword{functions}

[…]
