+++ R/productivity_levels.R
@@ -1,170 +1,0 @@
-#' Productivity Levels
-#'
-#' @export
-#'
-#' @param data matrix or tibble/data.frame (e.g. \code{world_trade_2017}).
-#' If the input is a matrix it must be a zero/one matrix with countries in rows
-#' and products in columns.
-#' If the input is a tibble/data.frame it must contain at least three columns
-#' with countries, products and values.
-#' @param c1 string to indicate the column that contains exporting
-#' countries in revealed_comparative_advantage (set to "country" by default)
-#' @param p1 string to indicate the column that contains exported products
-#' in revealed_comparative_advantage (set to "product" by default)
-#' @param v1 string to indicate the column that contains traded values in
-#' revealed_comparative_advantage (set to "value" by default)
-#' @param gdp vector or tibble/data.frame
-#' (e.g. \code{world_gdp_and_population_2017}).
-#' If the input is a vector it must be a numeric vector with optional names.
-#' If the input is a tibble/data.frame it must contain at least two columns with
-#' countries and values.
-#' @param c2 string to indicate the column that contains exporting
-#' countries in revealed_comparative_advantage (set to "country" by default)
-#' @param v2 string to indicate the column that contains traded values in
-#' revealed_comparative_advantage (set to "value" by default)
-#' @param tbl when set to TRUE the output will be a tibble instead of a
-#' matrix (default set to FALSE)
-#'
-#' @importFrom magrittr %>%
-#' @importFrom dplyr select group_by ungroup mutate summarise matches rename
-#'  pull inner_join
-#' @importFrom Matrix Matrix t rowSums colSums
-#' @importFrom stats setNames
-#' @importFrom rlang sym syms
-#'
-#' @examples
-#' ec_productivity_levels(
-#'   data = ec_trade_1962,
-#'   c1 = "country",
-#'   p1 = "product",
-#'   v1 = "value",
-#'   gdp = ec_gdp_pc_1962,
-#'   c2 = "country",
-#'   v2 = "value",
-#'   tbl = TRUE
-#' )
-#' @references
-#' For more information on prody and its applications see:
-#'
-#' \insertRef{atlas2014}{economiccomplexity}
-#'
-#' \insertRef{exportmatters2005}{economiccomplexity}
-#'
-#' @keywords functions
-
-ec_productivity_levels <- function(data = NULL,
-                                   c1 = "country",
-                                   p1 = "product",
-                                   v1 = "value",
-                                   gdp = NULL,
-                                   c2 = "country",
-                                   v2 = "value",
-                                   tbl = FALSE) {
-  # sanity checks ----
-  if (all(class(data) %in% c(
-    "data.frame", "matrix", "dgeMatrix",
-    "dsCMatrix", "dgCMatrix"
-  ) == FALSE)) {
-    stop("data must be a tibble/data.frame or a dense/sparse matrix")
-  }
-
-  if (all(class(gdp) %in% c("data.frame", "numeric") == FALSE)) {
-    stop("gdp must be a tibble/data.frame or numeric")
-  }
-
-  if (!is.character(c1) & !is.character(p1) &
-    !is.character(v1)) {
-    stop("c1, p1 and v1 must be character")
-  }
-
[…]
