+++ .github/workflows/check-full.yaml
@@ -32,7 +32,7 @@ jobs:
       - uses: r-lib/actions/setup-tinytex@master
         if: contains(matrix.config.args, 'no-manual') == false
 
-       - name: Install system dependencies
+      - name: Install system dependencies
         if: runner.os == 'Linux'
         run: |
           sudo apt-get update -y
