### 2019-11-24 bf0af8b
removed text equations
