+++ man/rca.Rd
@@ -35,11 +35,12 @@ A data frame.
 \insertCite{measuringcomplexity2015;textual}{economiccomplexity}
 }
 \details{
-Given a \eqn{C\times P} matrix (C for "countries"
-and P for "products") or an equivalent 3-columns data frame with exported
-values (X) as input, this function implements the equation:
-\deqn{RCA_{cp} = \frac{X_{cp}}{\sum_c X_{cp}} /
-\frac{\sum_p X_{cp}}{\sum_{c}\sum_{p} X_{cp}}}
+Given a data frame or matrix with exported values as input, this
+function uses the tidyverse to obtain grouped sums and compute RCA.
+
+Previous versions of this function used R base, but dplyr package
+proved to be consistently faster for this output across different platforms
+and specially in Windows.
 }
 \examples{
 rca(ec_trade_1962)
