+++ man/projections.Rd
@@ -50,14 +50,15 @@ A list with two data frames or matrices.
 from \insertCite{measuringcomplexity2015;textual}{economiccomplexity}
 }
 \details{
-Given a matrix of \eqn{C\times C} with country proximity values and
-a matrix of \eqn{P\times P} with product proximity values, this function
-obtains a simplified network in two steps:
-1) creates a network skeleton by appliying the minimum spanning tree
+Given a two data frames or matrices with proximity values, this
+function obtains a simplified network in two steps:
+
+1) Creates a network skeleton by appliying the minimum spanning tree
 algorithm from the igraph package, but multiplying proximities by minus one,
-so that the strongest links are considered for the network skeleton
-2) append additional links to the skeleton by aading the links with proximity
-values above a user-defined cutoff
+so that the strongest links are considered for the network skeleton.
+
+2) Appends additional links to the skeleton by aading the links with
+proximity values above a user-defined cutoff.
 }
 \examples{
 projections(
