+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: economiccomplexity
 Type: Package
-Title: Build bipartite networks such as countries and their exported products
+Title: Create bipartite networks such as countries and their exported products
 Version: 0.3
 Authors@R: c(
     person(given = "Mauricio",
@@ -28,14 +28,12 @@ Authors@R: c(
     person(family = "Open Trade Statistics",
       role = "dtc")
   )
-Description: A wrapper of different indices and networks commonly used in 
-  Economic Complexity to explore bipartite relations such as countries and their
-  exported products. These methods are also useful for different kind of 
-  relations such as countries and their spoken languages.
-  The functions within this package correspond to code implementations of the 
-  equations described in Hausmann, et al. (2005) <doi:10.3386/w11905>, 
-  Hausmann, et al. (2014) <doi:10.7551/mitpress/9647.001.0001>, and 
-  Mariani, et al. (2015) <doi:10.1140/epjb/e2015-60298-7>.
+Description: A wrapper of different computation methods to create bipartite 
+  networks such as countries and their exported products, and it can be used to
+  explore different kind of relations such as countries and their spoken 
+  languages. This package is implements the equations described in 
+  Mariani, et al. (2015) <doi:10.1140/epjb/e2015-60298-7> and 
+  Hausmann, et al. (2014) <doi:10.7551/mitpress/9647.001.0001>.
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
