+++ README.Rmd
@@ -22,15 +22,13 @@ knitr::opts_chunk$set(
 
 ## Description
 
-`economiccomplexity` is a package that implements different computation methods 
-to create bipartite networks such as countries and their exported products, and 
-it can be used to explore different kind of relations such as countries and 
-their spoken languages. 
-
-This package implements the equations described in Mariani, et al. (2015) 
-<doi:10.1140/epjb/e2015-60298-7> and Hausmann, et al. (2014) 
-<doi:10.7551/mitpress/9647.001.0001>.  
-
+A wrapper of different computation methods to create bipartite networks such as
+countries and their exported products, and different kind of relations such as
+countries and their spoken languages. This package implements the equations 
+described in 
+Mariani, et al. (2015) <doi:10.1140/epjb/e2015-60298-7> and 
+Hausmann, et al. (2014) <doi:10.7551/mitpress/9647.001.0001>.
+  
 ## Installation
 
 ```{r, eval = FALSE}
