### 2019-11-28 #1866 @pachamaltese
Hi
I pushed changes to the documentation as John asked.
