### 2020-01-13 e2ae709
all functions rewrote with base + matrix + igraph
