+++ R/projections.R
@@ -1,208 +1,122 @@
-#' @title Country-product bipartite network projections
+#' Projections of a bipartite network
 #'
-#' @description \code{rca} computes complexity indices following the definitions
-#' from \insertCite{measuringcomplexity2015;textual}{economiccomplexity}
+#' @description TBD
 #'
-#' @details Given a two data frames or matrices with proximity values, this
-#' function obtains a simplified network in two steps:
+#' @details TBD
 #'
-#' 1) Creates a network skeleton by appliying the minimum spanning tree
-#' algorithm from the igraph package, but multiplying proximities by minus one,
-#' so that the strongest links are considered for the network skeleton.
+#' @param proximity_source a data frame containing proximity
+#' values for the elements of set X (e.g. proximity_source from \code{proximity()})
+#' @param proximity_target a data frame containing proximity
+#' values for the elements of set Y (e.g. proximity_target from \code{proximity()})
+#' @param avg_links average number of connections for the projection of X
+#' (default set to 4)
+#' @param tolerance tolerance for proximity variation on each iteration until
+#' obtaining the desired average number of connections (default set to 0.05)
+#' @param compute which projection to compute. By default is "both" (both
+#' projections) but it can also be "source" or "target".
 #'
-#' 2) Appends additional links to the skeleton by aading the links with
-#' proximity values above a user-defined cutoff.
+#' @importFrom igraph graph_from_adjacency_matrix graph_from_data_frame mst as_data_frame degree delete.edges graph.difference E E<-
 #'
-#' @param proximity_c matrix or data frame with product proximity values
-#' @param proximity_p matrix or data frame with country proximity values
-#' @param cutoff_c all the links with a proximity below this value will be
-#' removed from the country projection (by default is 0.2)
-#' @param cutoff_p all the links with a proximity below this value will be
-#' removed from the product projection (by default is 0.4)
-#' @param compute "country", "product" or "both" (default) projections
-#' @param tbl TRUE (default) returns a data.frame and FALSE returns a matrix
-#' @param from_c column containing origin (applies only if proximity_c is a
-#' data.frame)
-#' @param to_c column containing destination (applies only if proximity_c is a
-#' data.frame)
-#' @param value_c column containing proximity values (applies only if
-#' proximity_c is a data.frame)
-#' @param from_p column containing origin (applies only if proximity_p is a
-#' data.frame)
-#' @param to_p column containing destination (applies only if proximity_p is a
-#' data.frame)
-#' @param value_p column containing proximity values (applies only if
-#' proximity_p is a data.frame)
-#'
+#' @examples
+#' projections(
+#'   proximity_source = economiccomplexity_output$proximity$proximity_source,
+#'   proximity_target = economiccomplexity_output$proximity$proximity_target
+#' )
 #' @references
-#' For more information about visualizing bipartite networks projections see:
-#'
-#' \insertRef{human2007}{economiccomplexity}
+#' For more information see:
 #'
 #' \insertRef{atlas2014}{economiccomplexity}
 #'
 #' and the references therein.
[…]
