+++ README.md
@@ -1,12 +1,14 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# Binet <img src="https://pachamaltese.github.io/binet/hexicon.svg" width=150 align="right" alt="sticker"/>
+# economiccomplexity <img src="https://pachamaltese.github.io/economiccomplexity/hexicon.svg" width=150 align="right" alt="sticker"/>
 
 [![R build
-status](https://github.com/pachamaltese/binet/workflows/R-CMD-check/badge.svg)](https://github.com/pachamaltese/binet/actions?workflow=R-CMD-check)
+status](https://github.com/pachamaltese/economiccomplexity/workflows/R-CMD-check/badge.svg)](https://github.com/pachamaltese/economiccomplexity/actions?workflow=R-CMD-check)
 [![Codecov test
-coverage](https://codecov.io/gh/pachamaltese/binet/branch/master/graph/badge.svg)](https://codecov.io/gh/pachamaltese/binet?branch=master)
+coverage](https://codecov.io/gh/pachamaltese/economiccomplexity/branch/master/graph/badge.svg)](https://codecov.io/gh/pachamaltese/economiccomplexity?branch=master)
+[![CRAN
+status](https://www.r-pkg.org/badges/version/economiccomplexity)](https://cran.r-project.org/package=economiccomplexity)
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -19,23 +21,23 @@ useful for different kind of relations such as The Product Space
 <doi:10.1073/pnas.0701361104>.
 
 The functions within this package correspond to code implementations of
-the equations described in Mariani, et al. (2015)
-<doi:10.1140/epjb/e2015-60298-7>, Hausmann, et al. (2014)
-<doi:10.7551/mitpress/9647.001.0001>, and Hausmann, et al. (2005)
+the equations described in Mariani, et al. (2015)
+<doi:10.1140/epjb/e2015-60298-7>, Hausmann, et al. (2014)
+<doi:10.7551/mitpress/9647.001.0001>, and Hausmann, et al. (2005)
 <doi:10.3386/w11905>.
 
 ## Installation
 
 ``` r
 # Install stable version from CRAN
-install.packages("binet")
+install.packages("economiccomplexity")
 
 # Install stable version from GitHub
-devtools::install_github("pachamaltese/binet")
+devtools::install_github("pachamaltese/economiccomplexity")
 ```
 
 ## Code of conduct
 
 Please note that this project is released with a [Contributor Code of
-Conduct](https://pachamaltese.github.io/binet/CODE_OF_CONDUCT.html). By
-participating in this project you agree to abide by its terms.
+Conduct](https://pachamaltese.github.io/economiccomplexity/CODE_OF_CONDUCT.html).
+By participating in this project you agree to abide by its terms.
