### 2020-01-13 c4b63c3
restored country-product syntax
