+++ R/proximity.R
@@ -6,22 +6,22 @@
 #'
 #' @param balassa_index a data frame (e.g. the output from
 #' \code{balassa_index()}).
-#' @param balassa_sum_source numeric vector or tibble/data.frame containing the
-#' Balassa sum for elements of set X (e.g. \code{balassa_sum_source} from
+#' @param balassa_sum_country numeric vector or tibble/data.frame containing the
+#' Balassa sum for elements of set X (e.g. \code{balassa_sum_country} from
 #' \code{complexity_measures()}).
-#' @param balassa_sum_target numeric vector or tibble/data.frame containing the
-#' Balassa sum for elements of set Y (e.g. \code{balassa_sum_target} from
+#' @param balassa_sum_product numeric vector or tibble/data.frame containing the
+#' Balassa sum for elements of set Y (e.g. \code{balassa_sum_product} from
 #' \code{complexity_measures()}).
 #' @param compute which proximity to compute. By default is "both" (both
-#' proximities) but it can also be "source" or "target".
+#' proximities) but it can also be "country" or "product".
 #'
 #' @importFrom Matrix t tril rowSums colSums
 #'
 #' @examples
 #' proximity(
 #'   balassa_index = economiccomplexity_output$balassa_index,
-#'   balassa_sum_source = economiccomplexity_output$complexity_measures$balassa_sum_source,
-#'   balassa_sum_target = economiccomplexity_output$complexity_measures$balassa_sum_target
+#'   balassa_sum_country = economiccomplexity_output$complexity_measures$balassa_sum_country,
+#'   balassa_sum_product = economiccomplexity_output$complexity_measures$balassa_sum_product
 #' )
 #' @references
 #' For more information see:
@@ -34,51 +34,51 @@
 #'
 #' @export
 
-proximity <- function(balassa_index, balassa_sum_source, balassa_sum_target, compute = "both") {
+proximity <- function(balassa_index, balassa_sum_country, balassa_sum_product, compute = "both") {
   # sanity checks ----
   if (class(balassa_index) != "dgCMatrix") {
     stop("'balassa_index' must be a dgCMatrix")
   }
 
-  if (class(balassa_sum_source) != "numeric" |
-    class(balassa_sum_target) != "numeric") {
-    stop("'balassa_sum_source' and 'balassa_sum_target' must be numeric")
+  if (class(balassa_sum_country) != "numeric" |
+    class(balassa_sum_product) != "numeric") {
+    stop("'balassa_sum_country' and 'balassa_sum_product' must be numeric")
   }
 
-  if (is.numeric(balassa_sum_source) & is.null(names(balassa_sum_source))) {
-    stop("'balassa_sum_source' cannot have NULL names")
+  if (is.numeric(balassa_sum_country) & is.null(names(balassa_sum_country))) {
+    stop("'balassa_sum_country' cannot have NULL names")
   }
 
-  if (is.numeric(balassa_sum_target) & is.null(names(balassa_sum_target))) {
-    stop("'balassa_sum_target' cannot have NULL names")
+  if (is.numeric(balassa_sum_product) & is.null(names(balassa_sum_product))) {
+    stop("'balassa_sum_product' cannot have NULL names")
   }
 
-  if (!any(compute %in% c("both", "source", "target"))) {
-    stop("'compute' must be 'both', 'source' or 'target'")
+  if (!any(compute %in% c("both", "country", "product"))) {
+    stop("'compute' must be 'both', 'country' or 'product'")
   }
 
   # compute proximity matrices ----
 
   if (compute == "both") {
-    compute2 <- c("source", "target")
[…]
