+++ man/productivity_levels.Rd
@@ -1,0 +1,47 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/productivity_levels.R
+\name{productivity_levels}
+\alias{productivity_levels}
+\title{Productivity Levels}
+\usage{
+productivity_levels(
+  d1,
+  d2,
+  country = "country",
+  product = "product",
+  value = "value"
+)
+}
+\arguments{
+\item{d1}{foo}
+
+\item{d2}{bar}
+
+\item{country}{country}
+
+\item{product}{product}
+
+\item{value}{value}
+}
+\value{
+
+}
+\description{
+foo
+}
+\details{
+bar
+}
+\examples{
+set.seed(1810)
+d1 <- galactic_federation
+d2 <- setNames(rnorm(1:nrow(galactic_federation), 1000, 200), rownames(galactic_federation))
+pl <- productivity_levels(d1, d2)
+
+}
+\references{
+For more information on prody and its applications see:
+
+\insertRef{exportmatters2005}{economiccomplexity}
+}
+\keyword{functions}
