### 2020-01-13 f818634
trying to optimize complexity outlook
