### 2020-01-13 5aa6802
efficient version of COCG
