### 2020-01-13 deac1b5
sketch of tidy COCG
