+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(balassa_index)
 export(complexity_measures)
+export(countries_position)
 export(productivity_levels)
 export(projections)
 export(proximity)
