### 2020-01-14 8626bb4
using tcrossprod for coi/cog
