+++ R/countries_position.R
@@ -10,7 +10,7 @@
 #' @param proximity_product (Type: matrix or dgCMatrix) foo.
 #' @param complexity_index_product (Type: numeric) foobar.
 #'
-#' @importFrom Matrix Matrix rowSums colSums t
+#' @importFrom Matrix Matrix tcrossprod rowSums colSums t
 #'
 #' @examples
 #' countries_position(
@@ -31,24 +31,31 @@
 #' @export
 
 countries_position <- function(balassa_index, proximity_product, complexity_index_product) {
-  dcp <- t(t((1 - balassa_index) %*% proximity_product) / rowSums(proximity_product))
-
-  coc <- ((1 - dcp) * (1 - balassa_index)) %*% complexity_index_product
-
-  cogc <- Matrix(0, nrow = nrow(balassa_index), ncol = ncol(balassa_index), sparse = TRUE,
-                 dimnames = list(rownames(balassa_index), colnames(balassa_index)))
-
-  for (i in 1:nrow(balassa_index)) {
-    p1 <- t((t(proximity_product / colSums(proximity_product)) * (1 - balassa_index[i, ])) %*% complexity_index_product)
-    p2 <- dcp[i, ] * complexity_index_product
-    cogc[i, ] <- p1 - p2
+  # sanity checks ----
+  if (class(balassa_index) != "dgCMatrix") {
+    stop("'balassa_index' must be a dgCMatrix")
   }
+
+  if (class(proximity_product) != "dgCMatrix") {
+    stop("'proximity_product' must be a dgCMatrix")
+  }
+
+  if (class(complexity_index_product) != "dgCMatrix") {
+    stop("'complexity_index_product' must be a dgCMatrix")
+  }
+
+  # compute matrices ----
+  density <- tcrossprod(balassa_index, proximity_product / rowSums(proximity_product))
+
+  coi <- rowSums(t(t(density * (1 - balassa_index)) * complexity_index_product))
+
+  cog <- (1 - balassa_index) * tcrossprod((1 - balassa_index),
+    t(proximity_product * (complexity_index_product / rowSums(proximity_product))))
 
   return(
     list(
-      proximity_distance = dcp,
-      complexity_outlook = coc,
-      complexity_outlook_gain = cogc
+      complexity_outlook_index = coi,
+      complexity_outlook_gain = cog
     )
   )
 }
