### 2020-01-14 f760d4a
crossprod optimization
