+++ man/complexity_outlook.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/countries_position.R
-\name{countries_position}
-\alias{countries_position}
-\title{Countries Position}
+% Please edit documentation in R/complexity_outlook.R
+\name{complexity_outlook}
+\alias{complexity_outlook}
+\title{Complexity Outlook}
 \usage{
-countries_position(balassa_index, proximity_product, complexity_index_product)
+complexity_outlook(balassa_index, proximity_product, complexity_index_product)
 }
 \arguments{
 \item{balassa_index}{(Type: matrix or dgCMatrix) foo}
@@ -23,7 +23,7 @@ foo
 bar
 }
 \examples{
-countries_position(
+complexity_outlook(
  economiccomplexity_output$balassa_index,
  economiccomplexity_output$proximity$proximity_product,
  economiccomplexity_output$complexity_measures$complexity_index_product
