+++ NAMESPACE
@@ -2,15 +2,17 @@
 
 export(balassa_index)
 export(complexity_measures)
-export(countries_position)
+export(complexity_outlook)
 export(productivity_levels)
 export(projections)
 export(proximity)
 importFrom(Matrix,Matrix)
 importFrom(Matrix,colSums)
+importFrom(Matrix,crossprod)
 importFrom(Matrix,rowSums)
 importFrom(Matrix,sparseMatrix)
 importFrom(Matrix,t)
+importFrom(Matrix,tcrossprod)
 importFrom(Matrix,tril)
 importFrom(igraph,"E<-")
 importFrom(igraph,E)
