+++ R/balassa_index.R
@@ -7,12 +7,9 @@
 #' \insertCite{measuringcomplexity2015}{economiccomplexity} to obtain a metric for
 #' specialisation. In the context of international trade, if the Balassa Index
 #' for a country-product pair is more than 1, it means that country is
-#' specialized in that product.
-#'
-#' If the input for this function is a data.frame instead of a matrix, the
-#' function shall aggregate the data and convert the input to a matrix. You
-#' don't need to aggregate the data beforehand if, for example, you pass a
-#' data.frame containing origin, destination, product and exports value.
+#' specialized in that product. If the input for this function is a data.frame
+#' instead of a matrix, the function shall aggregate the data and convert the
+#' input to a matrix.
 #'
 #' @return A matrix with the Balassa Index.
 #'
@@ -24,12 +21,12 @@
 #' @param cutoff (Type: numeric) the cutoff to use for discretization.
 #' By default this is set to \code{1}.
 #' @param country (Type: character) the column with the countries.
-#' By default this is set to \code{"country"}. Applies only if the input is a data.frame.
+#' By default this is set to \code{"country"}. Used only if the input is a data.frame.
 #' @param product (Type: character) the column with the products.
-#' By default this is set to \code{"product"}. Applies only if the input is a data.frame.
+#' By default this is set to \code{"product"}. Used only if the input is a data.frame.
 #' @param value  (Type: character) the column with the metric for
 #' country-product pairs.
-#' By default this is set to \code{"value"}. Applies only if the input is a data.frame.
+#' By default this is set to \code{"value"}. Used only if the input is a data.frame.
 #'
 #' @importFrom Matrix Matrix rowSums colSums t
 #'
