### 2020-01-16 0c9c76a
simplified projections + tidy documentation
