+++ NAMESPACE
@@ -14,12 +14,12 @@ importFrom(Matrix,sparseMatrix)
 importFrom(Matrix,t)
 importFrom(Matrix,tcrossprod)
 importFrom(igraph,"E<-")
-importFrom(igraph,E)
+importFrom(igraph,as_data_frame)
 importFrom(igraph,degree)
 importFrom(igraph,delete.edges)
 importFrom(igraph,graph.difference)
-importFrom(igraph,graph.union)
 importFrom(igraph,graph_from_adjacency_matrix)
+importFrom(igraph,graph_from_data_frame)
 importFrom(igraph,mst)
 importFrom(stats,aggregate)
 importFrom(stats,cor)
