### 2020-01-16 4c2b845
documented functions
