+++ R/productivity_levels.R
@@ -1,25 +1,40 @@
 #' Productivity Levels
 #'
-#' @description foo
+#' @description \code{productivity_levels()} computes EXPY and PRODY measures.
 #'
-#' @details bar
+#' @details The current implementation follows
+#' \insertCite{exportmatters2005}{economiccomplexity} to obtain the
+#' productivity and income levels associated to exports.
 #'
-#' @return foobar
+#' @return A list of two named numeric vectors.
 #'
-#' @param d1 foo
-#' @param d2 bar
-#' @param country country
-#' @param product product
-#' @param value value foobar
+#' @param data_exp (Type: data.frame, matrix or dgCMatrix) a dataset containing
+#' countries, products and exported values.
+#' @param data_gdp (Type: data.frame, matrix or dgCMatrix) a dataset
+#' containing countries and per-capita GDP values.
+#' @param country (Type: character) the column with the countries.
+#' By default this is set to \code{"country"}. Used only if the input is a data.frame.
+#' @param product (Type: character) the column with the products.
+#' By default this is set to \code{"product"}. Used only if the input is a data.frame.
+#' @param value  (Type: character) the column with the metric for
+#' country-product pairs.
+#' By default this is set to \code{"value"}. Used only if the input is a data.frame.
 #'
 #' @importFrom Matrix Matrix rowSums colSums t crossprod
 #' @importFrom stats setNames
 #'
 #' @examples
+#' data_exp <- galactic_federation
+#'
+#' # Since I don't have a gdp per capita dataset for the Galactic Federation,
+#' # I'll create simulated data
 #' set.seed(1810)
-#' d1 <- galactic_federation
-#' d2 <- setNames(rnorm(1:nrow(galactic_federation), 1000, 200), rownames(galactic_federation))
-#' pl <- productivity_levels(d1, d2)
+#' data_gdp <- setNames(
+#'  rnorm(1:nrow(galactic_federation), 1000, 200),
+#'  rownames(galactic_federation)
+#' )
+#'
+#' pl <- productivity_levels(data_exp, data_gdp)
 #'
 #' @references
 #' For more information on prody and its applications see:
@@ -30,50 +45,54 @@
 #'
 #' @export
 
-productivity_levels <- function(d1, d2,
+productivity_levels <- function(data_exp, data_gdp,
                                 country = "country", product = "product", value = "value") {
   # sanity checks ----
-  if (all(class(d1) %in% c("data.frame", "matrix", "dgCMatrix") == FALSE)) {
-    stop("'d1' must be a data.frame, matrix or dgCMatrix")
+  if (all(class(data_exp) %in% c("data.frame", "matrix", "dgCMatrix") == FALSE)) {
+    stop("'data_exp' must be a data.frame, matrix or dgCMatrix")
   }
 
-  if (all(class(d2) %in% c("data.frame", "numeric") == FALSE)) {
-    stop("'d2' must be a data.frame or numeric")
+  if (all(class(data_gdp) %in% c("data.frame", "numeric") == FALSE)) {
+    stop("'data_gdp' must be a data.frame or numeric")
   }
 
-  # tidy input data d1 ----
-  if (any(class(d1) %in% "data.frame")) {
-    d1 <- country_product_aggregation(d1, country, product, value)
-    d1 <- dataframe_to_matrix(d1, country, product, value)
+  # tidy input data data_exp ----
+  if (any(class(data_exp) %in% "data.frame")) {
[…]
