### 2020-01-17 c2b96f6
united graphs without rbind step
