+++ NAMESPACE
@@ -15,13 +15,13 @@ importFrom(Matrix,t)
 importFrom(Matrix,tcrossprod)
 importFrom(igraph,"E<-")
 importFrom(igraph,E)
-importFrom(igraph,as_data_frame)
 importFrom(igraph,degree)
 importFrom(igraph,delete.edges)
 importFrom(igraph,graph.difference)
+importFrom(igraph,graph.union)
 importFrom(igraph,graph_from_adjacency_matrix)
-importFrom(igraph,graph_from_data_frame)
 importFrom(igraph,mst)
+importFrom(igraph,remove.edge.attribute)
 importFrom(stats,aggregate)
 importFrom(stats,cor)
 importFrom(stats,sd)
