### 2020-01-17 #1866 @pachamaltese
Hi
Just to inform that I rewrote all functions toward easier use and speed
