+++ dev/product_space_replication.R
@@ -1,59 +1,45 @@
-library(binet)
+library(economiccomplexity)
+library(dplyr)
+library(igraph)
+library(ggraph)
 
-trade <- readRDS("~/github/binet/dev/world_trade_avg_1998_to_2000.rds")
+trade <- readRDS("~/github/economiccomplexity/dev/world_trade_avg_1998_to_2000.rds")
 
 bi <- balassa_index(
   data = trade,
-  source = "reporter_iso",
-  target = "product_code",
+  country = "reporter_iso",
+  product = "product_code",
   value = "trade_value_usd"
 )
 
 com_fit <- complexity_measures(balassa_index = bi)
 
 pro <- proximity(
-  balassa_index = bi,
-  balassa_sum_source = com_fit$balassa_sum_source,
-  balassa_sum_target = com_fit$balassa_sum_target
+  balassa_index = bi
 )
 
 net <- projections(
-  proximity_source = pro$proximity_source,
-  proximity_target = pro$proximity_target,
+  proximity_country = pro$proximity_country,
+  proximity_product = pro$proximity_product,
   tolerance = 0.05
 )
-
-aggregated_countries <- trade %>%
-  group_by(reporter_iso) %>%
-  summarise(trade_value_usd = sum(trade_value_usd))
 
 aggregated_products <- trade %>%
   group_by(product_code) %>%
   summarise(trade_value_usd = sum(trade_value_usd))
 
+aggregated_products <- setNames(aggregated_products$trade_value_usd, aggregated_products$product_code)
+V(net$network_product)$size <- aggregated_products[match(V(net$network_product)$name, names(aggregated_products))]
+
 set.seed(200100)
 
-g_countries <- net$network_source %>%
-  graph_from_data_frame(vertices = aggregated_countries, directed = F) %>%
-  ggraph(layout = "kk") +
-  geom_edge_link(aes(edge_alpha = value, edge_width = value),
-                 edge_colour = "#a8a8a8") +
-  geom_node_point(aes(size = trade_value_usd), color = "darkslategray4") +
-  scale_size(range = c(2,10)) +
-  geom_node_text(aes(label = name), vjust = 2.2) +
-  ggtitle("Proximity Based Network Projection for Planets") +
-  theme_void()
-
-g_products <- net$network_target %>%
-  graph_from_data_frame(vertices = aggregated_products, directed = F) %>%
-  ggraph(layout = "kk") +
-  geom_edge_link(aes(edge_alpha = value, edge_width = value),
-                 edge_colour = "#a8a8a8") +
-  geom_node_point(color = "darkslategray4", size = 4) +
-  geom_node_point(aes(size = trade_value_usd), color = "darkslategray4") +
-  scale_size(range = c(2,10)) +
-  geom_node_text(aes(label = name), vjust = 2.2) +
+g_products <- net$network_product %>%
+  ggraph(layout = "fr") +
+  # geom_edge_link(aes(edge_width = weight), edge_colour = "#a8a8a8") +
+  geom_edge_link(edge_colour = "#888888") +
+  geom_node_point(aes(size = size), color = "#86494d") +
+  # geom_node_text(aes(label = name), vjust = 2.2) +
   ggtitle("Proximity Based Network Projection for Products") +
   theme_void()
 
-data.table::fwrite(net$network_target, "dev/product_space_replication.csv")
+g_products
