+++ R/productivity_levels.R
@@ -26,7 +26,7 @@
 #' @examples
 #' data_exp <- galactic_federation
 #'
-#' # Since I don't have a gdp per capita dataset for the Galactic Federation,
+#' # Since I don't have a per-capita GDP dataset for the Galactic Federation,
 #' # I'll create simulated data
 #' set.seed(1810)
 #' data_gdp <- setNames(
@@ -34,7 +34,7 @@
 #'  rownames(galactic_federation)
 #' )
 #'
-#' pl <- productivity_levels(data_exp, data_gdp)
+#' productivity_levels(data_exp, data_gdp)
 #'
 #' @references
 #' For more information on prody and its applications see:
