### 2020-01-18 efec16e
adds knitr to gh actions
