### 2020-01-19 43576e4
removes galactic federation data and examples
