+++ tests/testthat/test-productivity_levels.R
@@ -1,111 +1,98 @@
 test_that("productivity_levels works with a sparse matrix + named vector", {
-  set.seed(1810)
-
-  pl <- productivity_levels(
-    galactic_federation,
-    setNames(rnorm(1:nrow(galactic_federation), 1000, 200), rownames(galactic_federation))
+  pl <- expect_warning(
+    productivity_levels(
+      world_trade_avg_1998_to_2000,
+      world_gdp_avg_1998_to_2000
+    )
   )
 
   expect_is(pl, "list")
-  expect_equal(length(pl$productivity_level_country), 9)
-  expect_equal(length(pl$productivity_level_product), 12)
+  expect_equal(length(pl$productivity_level_country), 185)
+  expect_equal(length(pl$productivity_level_product), 785)
 })
 
 test_that("productivity_levels works with a dense matrix + named vector", {
-  set.seed(1810)
-
-  pl <- productivity_levels(
-    as.matrix(galactic_federation),
-    setNames(rnorm(1:nrow(galactic_federation), 1000, 200), rownames(galactic_federation))
+  pl <- expect_warning(
+    productivity_levels(
+      as.matrix(world_trade_avg_1998_to_2000),
+      world_gdp_avg_1998_to_2000
+    )
   )
 
   expect_is(pl, "list")
-  expect_equal(length(pl$productivity_level_country), 9)
-  expect_equal(length(pl$productivity_level_product), 12)
+  expect_equal(length(pl$productivity_level_country), 185)
+  expect_equal(length(pl$productivity_level_product), 785)
 })
 
 test_that("productivity_levels works with a data frame + named vector", {
-  set.seed(1810)
+  wt <- Matrix::as.matrix(world_trade_avg_1998_to_2000)
+  wt <- as.data.frame(as.table(wt), stringsAsFactors = FALSE)
 
-  gf <- Matrix::as.matrix(galactic_federation)
-  gf <- as.data.frame(as.table(gf), stringsAsFactors = FALSE)
-
-  gf2 <- setNames(rnorm(1:nrow(galactic_federation), 1000, 200), rownames(galactic_federation))
-
-  pl <- productivity_levels(
-    galactic_federation,
-    gf2
+  pl <- expect_warning(
+    productivity_levels(
+      world_trade_avg_1998_to_2000,
+      world_gdp_avg_1998_to_2000
+    )
   )
 
-  pl2 <- productivity_levels(
-    gf,
-    gf2,
-    country = "Var1", product = "Var2", value = "Freq"
+  pl2 <- expect_warning(
+    productivity_levels(
+      wt,
+      world_gdp_avg_1998_to_2000,
+      country = "Var1", product = "Var2", value = "Freq"
+    )
   )
 
   expect_is(pl2, "list")
   expect_equal(pl$productivity_level_country, pl2$productivity_level_country)
   expect_equal(pl$productivity_level_product, pl2$productivity_level_product)
-  expect_equal(length(pl$productivity_level_country), 9)
-  expect_equal(length(pl$productivity_level_product), 12)
+  expect_equal(length(pl$productivity_level_country), 185)
+  expect_equal(length(pl$productivity_level_product), 785)
 })
 
 test_that("productivity_levels works with a data frame + data frame", {
   set.seed(1810)
 
-  gf <- Matrix::as.matrix(galactic_federation)
-  gf <- as.data.frame(as.table(gf), stringsAsFactors = FALSE)
+  wt <- Matrix::as.matrix(world_trade_avg_1998_to_2000)
+  wt <- as.data.frame(as.table(wt), stringsAsFactors = FALSE)
 
[…]
