+++ man/rca.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/balassa_index.R
+\name{rca}
+\alias{rca}
+\title{Revealed Comparative Advantage}
+\usage{
+rca(...)
+}
+\arguments{
+\item{...}{old parameters}
+}
+\description{
+\code{balassa_index()} replaces this function
+}
