+++ NAMESPACE
@@ -6,6 +6,7 @@ export(complexity_outlook)
 export(productivity_levels)
 export(projections)
 export(proximity)
+export(rca)
 importFrom(Matrix,Matrix)
 importFrom(Matrix,colSums)
 importFrom(Matrix,crossprod)
