### 2020-01-22 64a787d
updated readme + 100% coverage
