+++ README.md
@@ -15,13 +15,12 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 [![Lifecycle:
 maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 
-A consistent tool to explore bipartite relations. These methods are
-useful for different kind of relations such as The Product Space
-<doi:10.7551/mitpress/9647.001.0001> and The Human Disease Network
-<doi:10.1073/pnas.0701361104>.
+A wrapper of different methods from Linear Algebra for the equations
+introduced in The Atlas of Economic Complexity and related literature.
+This package provides standard matrix and graph output that can be used
+seamlessly with other packages.
 
-The functions within this package correspond to code implementations of
-the equations described in Mariani, et al. (2015)
+The references are Mariani, et al. (2015)
 <doi:10.1140/epjb/e2015-60298-7>, Hausmann, et al. (2014)
 <doi:10.7551/mitpress/9647.001.0001>, and Hausmann, et al. (2005)
 <doi:10.3386/w11905>.

[…]
