### 2020-01-27 515520d
complexity measures divided into smaller funs (rstudio conf tips)
