+++ man/eigenvalues_method.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/complexity_measures.R
+\name{eigenvalues_method}
+\alias{eigenvalues_method}
+\title{Eigenvalues Method}
+\usage{
+eigenvalues_method(balassa_index, iterations)
+}
+\arguments{
+\item{...}{params inherited from \code{complexity_measures()}}
+}
+\description{
+Eigenvalues Method
+}
+\keyword{internal}
