+++ man/fitness_method.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/complexity_measures.R
+\name{fitness_method}
+\alias{fitness_method}
+\title{Fitness Method}
+\usage{
+fitness_method(balassa_index, iterations, extremality)
+}
+\arguments{
+\item{...}{params inherited from \code{complexity_measures()}}
+}
+\description{
+Fitness Method
+}
+\keyword{internal}
