### 2020-01-27 be573b9
add @params to internal funs + reduced examples time
