+++ man/projections.Rd
@@ -48,7 +48,9 @@ returned when is not possible to return the desired network.
 \examples{
 net <- projections(
  economiccomplexity_output$proximity$proximity_country,
- economiccomplexity_output$proximity$proximity_product
+ economiccomplexity_output$proximity$proximity_product,
+ avg_links = 10,
+ tolerance = 0.1
 )
 
 # partial view of projections
