### 2020-02-04 #1866 @mpadge
Thanks @oliviaguest, it'll be a pleasure to review this exciting looking package. I'll try to get it done this week (<= 7 Feb 2020); if not I'll commit myself now to ensuring it's done by end of the week after (<= 14 Feb 2020). I trust that's okay
