### 2020-02-14 #1866 @mpadge
just done. All good to go, except for one (very) minor comment i made regarding the `paper.md`, which could perhaps be improved prior to official publication by one final edit.
