### 2020-02-14 pachamaltese/economiccomplexity#12 @JohnCoene
Reference Paper
Hi Mauricio,

In the paper I read that the package depends on the tidyverse but in the DESCRIPTION I see no tidyverse packages.

I could be missing something.
