### 2020-02-14 #1866 @JohnCoene
@oliviaguest A wee tiny thing with the paper, an opened an issue [#12](https://github.com/pachamaltese/economiccomplexity/issues/12). When this is addressed it's all good on my end.
