### 2020-02-14 pachamaltese/economiccomplexity#12 @pachamaltese
Sorry! It was dplyr based but now it's Matrix based. I'm fixing it on
Saturday

On Fri, Feb 14, 2020, 8:21 AM John (JP) Coene <notifications@github.com>
wrote:

> Hi Mauricio,
>
> In the paper I read that the package depends on the tidyverse but in the
> DESCRIPTION I see no tidyverse packages.
>
> I could be missing something.
>
