+++ inst/REFERENCES.bib
@@ -149,10 +149,12 @@
     url = {https://CRAN.R-project.org/package=Matrix}
 }
 
-@Manual{igraph,
-   title = {Matrix: Sparse and Dense Matrix Classes and Methods},
-   author = {Douglas Bates and Martin Maechler},
-   year = {2019},
-   note = {R package version 1.2-18},
-   url = {https://CRAN.R-project.org/package=Matrix}
+@Article{igraph,
+    title = {The igraph software package for complex network research},
+    author = {Gabor Csardi and Tamas Nepusz},
+    journal = {InterJournal},
+    volume = {Complex Systems},
+    pages = {1695},
+    year = {2006},
+    url = {http://igraph.org}
 }
