### 2020-02-16 pachamaltese/economiccomplexity#12 @JohnCoene
Thanks @pachamaltese, this seems solves now. If you close this I'm good to approve the review on my side: sorry, I know this took some time.
