### 2020-02-16 #1866 @pachamaltese
> [ point_right Check article proof page_facing_up point_left ](https://github.com/openjournals/joss-papers/blob/joss.01866/joss.01866/10.21105.joss.01866.pdf)

@JohnCoene this closed #12
