+++ inst/REFERENCES.bib
@@ -74,7 +74,8 @@
   volume={2},
   pages={723},
   year={2012},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/srep00723}
 }
 
 @book{atlas2014,
@@ -126,7 +127,8 @@
   volume={5},
   pages={8182},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/srep08182}
 }
 
 @article{linking2017,
