### 2020-02-17 #1866 @pachamaltese
> box checked and all wholeheartedly approved from my side

Dr. Padgham
I asked a native speaker to read the text and make some parts be more "fluid".
Thanks a lot for the useful suggestions.
