### 2020-02-17 #1866 @oliviaguest
@pachamaltese  Sorry zenodo! Here are the links, use whatever works best for you:
* https://figshare.com/
* https://zenodo.org/

I'm sure others exist too! 😅
