### 2020-02-17 #1866 @pachamaltese
> @pachamaltese Sorry zenodo! Here are the links, use whatever works best for you:
>
> * https://figshare.com/
> * https://zenodo.org/
>
> I'm sure others exist too! 😅

sure! now it makes more sense, I was creating an entry in zotero
