### 2020-02-17 #1866 @pachamaltese
> Thank you! @pachamaltese can you check the final proof please? 🎉

Dear Dr. @oliviaguest
I read ok in my opinion. The only thing I dislike is a reference that created a new page for just 3 lines. Besides it, I'm ok with the result.
