### 2020-02-17 b83b7c2
eigenvalues short description
