### 2020-02-17 #1866 @pachamaltese
Dear Dr. @oliviaguest

I changed the citations () and I re-added a paragraph that I had removed before.

I'm also tagging @JohnCoene and @mpadge on your behalf.

Best,
