### 2020-02-18 #1866 @pachamaltese
@oliviaguest
It looks ok imo and I'm waiting for the final public version

On Tue, Feb 18, 2020, 9:09 AM Mauricio Vargas <mauriciovargas@ug.uchile.cl>
wrote:

> Thanks a lot!!
>
> On Tue, Feb 18, 2020, 8:40 AM Olivia Guest <notifications@github.com>
> wrote:
>
>> It doesn't matter. If you like @pachamaltese
>> <https://github.com/pachamaltese>, we are good to do. 🎉
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/1866?email_source=notifications&email_token=ACM7UOPTFA5XN64LYEDSUGDRDPCI5A5CNFSM4JJGHPP2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEMBVCHA#issuecomment-587419932>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ACM7UOIFRVMRPMXKNUQAVHTRDPCI5ANCNFSM4JJGHPPQ>
>> .
>>
>
