### 2020-02-19 #1866 @oliviaguest
@pachamaltese no, we just wait for a @openjournals/joss-eics for the final step. Thank you.
