### 2020-02-19 882fae2
Update paper.md
Some minor copy-editing, removing "see" before some citations and correcting one citation method
