+++ inst/paper.md
@@ -30,14 +30,14 @@ products that they export.
 Seminal papers in economic complexity are: @productspace2007, which introduces 
 graphs to explore trade diversification and export opportunities; and 
 @buildingblocks2009, which develops dedicated metrics of complexity. These two
-articles are expanded and interpreted in @atlas2014, whose typesetted equations
+articles are expanded and interpreted by @atlas2014, whose typesetted equations
 were translated to code in the R package **economiccomplexity**.
 
 This R package provides different methods to compute complexity metrics that ease
 access to this line of research for social scientists. `economiccomplexity::`
 depends on `Matrix::` and `igraph::`, and
-follows the same design philosophy, grammar, and data structures from it (See
-@matrix and @igraph).
+follows the same design philosophy, grammar, and data structures from it
+[@matrix; @igraph].
 
 Recent articles such as @linking2017 introduce the question whereas a country's
 mix of products could predict its pattern of diversification and income
@@ -69,28 +69,28 @@ minimum standard when full independent replication of a study is not possible,
 and that becomes even more important to evaluate scientific claims in studies
 with public policy implications. A tenet of the scientific method holds that
 every research finding should be reproducible before it becomes accepted as a
-genuine contribution to human knowledge (@computing2009).
+genuine contribution to human knowledge [@computing2009].
 
 Complexity methods are also used outside its original research area. In
 particular, the fitness method from @metrics2012 has been used in ecology to
-study species interaction (see @ranking2015) and, more directly related, to
-study the scientific competitiveness of nations (See @scientific2014).
+study species interaction [@ranking2015] and, more directly related, to
+study the scientific competitiveness of nations [@scientific2014].
 
 The application widespread of economic complexity should be considered with 
 caution. Any theoretical or computational implementation shall be far
 from becoming a keystone if reproducibility is ignored. The extent to which code
 in computational research would build with reasonable effort is lower than 20%
-(See @measuring2014). A desirable growth pattern should focus on reproducibility,
+[@measuring2014]. A desirable growth pattern should focus on reproducibility,
 and I hope this package means a contribution to transparent research
 practices.
 
 @atlas2014 proposes the equivalency between the reflections and 
-the eigenvalues methods that they present.My code on GitHub
+the eigenvalues methods that they present. My code on GitHub
 checks for the sign of the correlation for the vector output of the two methods and 
 corrects the eigenvalues output when necessary. What motivated those additional
 steps is that, what I initially tought it was an error, turned out to be a 
 particular case that emerges with some datasets. Shortly after implementing that,
 both Python and Stata users started sending me emails when they, independently,
-found the same problem and  then search engines lead them to my package.
+found the same problem and  then search engines led them to my package.
 
 # References
