### 2020-02-19 #1866 @kyleniemeyer
Hi @pachamaltese, I did some minor copy-editing in this PR: https://github.com/pachamaltese/economiccomplexity/pull/14

Could you merge this and let me know when you have?
